# Added a benchmark for map overlap